Add Sqlite backend for Output Manager Service
Add a Sqlite storage backed implementation for the Output Manager Service and provide a test that tests both the in memory and sqlite backends using the same logic.

The backend is not included as part of the default Wallet build yet as the In-Memory DB is more consistent for initial mobile app dev but can be enabled with a single line change when we are ready.