Rework Difficulty for multiple PoW
*Note:* The Reorg logic has regressed due to this change, but I'm
pushing this PR now becuase it is blocking other development.

With Tari deciding to use hybrid mining (Merge mining + standalone),
we can no longer have a naked `difficulty` field in the block header.

To make extending the PoW to three or more algos easier, the difficulty
field(s) are moved into a separate `ProofOfWork` struct that contains
the separate accumulated difficulties for the various algos as well as
any metadata that is required for validation. For example, in Monero
merge mining, you need the Monero header and the RandomX seed key.

Blake mining doesn't need any additional metadata.

The concept of `total_accumulated_difficulty` is no longer stateless
because we now have to calculate a scaling factor between competing
POW algorithms, which in turn uses historical mining info.

This makes it impractical to use `total_difficulty` anywhere, so it's
been removed from:
* Chain metadata
* Mining algorithms
* Blockchain DB validation (temporary. Once ValidationPipeline is merged
we can use state to calculate and validate copeting forks based on total
accumulated work)

Also changed:
* Protobuf defintiions updated where appropriate.
* ProofOfWork trait dropped in favour of a specialised struct
* BlockHeader hash fixed to include nonce and PoW metadata, since it
  should not be malleable.
* `BlockHeader::from_previous` will now correctly update the
  `self.pow.accumulated_nnn_difficulty` fields from the previous header.
* The re-org logic on blockchaindb has regressed since total difficulty
  is no longer available with out state. I've blocked out the test
  for it, and put in some stubs with the basis of a suggested refactor.
  This should be re-implemented as soon as possible.
* Future note: All validation logic will soon be moving into one or
  more `ValidationPipeline`s, which can hold state and consensus rules
  in particular. This will clean up an simplify things. Currently
  validation is spread out all over the place.