Validation Pipeline
The idea of ValidationPipeline is to provide a unified approach to
validation; and one that makes testing easier as we approach full
set of consensus rules.

A validation pipeline contains a set of validators that must all
pass before the pipeline passes.

The pipeline can maintain state (such as a BlockchainDB instance)
in order to help with validation.

There are also mock validators provided so that we can keep unit tests
of different parts of the consensus model separate, without
having to rewrite tests every time we add another consesnsus rule
in development.

This also lets us keep all validation in one place, making
maintainability easier.