Blockchain database validators
We provde a set of `Validators` to `BlockchainDatabase` so that
it can test whether new blocks are valid.

We use a dedicated Validator trait rather than embedding functions
into the DB because it makes testing easier; this PR being a case
in point. Each time we update the validation rules, all the unit
tests for other aspects of the blockchain that make use of a
test DB would need to be updated. For this reason, `MockValidator`
is used everywhere to allow the current test suite to pass.

The actual Base Node executable itself uses a correct set of
Validators.

The validators are still missing some steps, which have been
marked in the code, with correspondinng issues assigned to
the TODOs.