Fix the usage of TempDir
We were using TempDir incorrectly. We were essentially using it to generate a random path for us, the letting it delete that directory but then remake the directly ourselves but never clean it up.

There was a line in the comms_setup that did this with a silent error which I have removed and then fixed all the usages of tempdir to be used correctly so the resulting directory is cleaned at the end of the test.