Add Coinbase Transaction Builder
The coinbase transaction hasn't been properly addressed until now.
There were several pieces of code that provided CB TXs in various
places in the codebase.

The CB TX is a simple but non-trivial implementation and so a builder
pattern is used.

The builder unifies all the code for generating a CB TX in one place.

The builder lives in the `transactions` crate and so does not know
about a block. Another PR will provide a `from_block` type function
in `tari_core` that will make use of the builder to ergonomically
construct a CB TX from information in a block.