Provide async callback for when Transaction service needs to perform peer discovery
When a transaction is sent to a peer that is not known the comms stack will perform a Discovery process which can take minutes. Do not block the client using the wallet library we now return immediately telling the client that the Discovery process is underway and then when it is completed we will use a call-back and event bus to let the client know that the Discovery has been completed (successfully or not) asynchronously.

If the discovery was unsuccessful the pending transaction is cancelled.