Add support for managing coinbase transactions
This PR add the functionality for a Miner to request the spending key for a coinabase output that they want to mine. The key and amount of the coinbase is stored as a Pending Incoming Transaction and will reflect in the pending balance of the wallet.

If the mining is successful the Miner will supply the completed transaction to the Transaction Service. 

If mining was unsuccessful the coinbasetransaction should be cancelled.