Allow Liveness Service to monitor Liveness stats for specified nodes
This PR adds the ability to register NodeIDs to be monitored with the Liveness service which will then monitor their Liveness using periodic Pings and maintain stats regarding the Liveness. This is to enable a wallet to monitor the status of the Base Nodes it wants to use to interface with the base layer.