Modify MutableMmr calls to use new constructor
All uses of MutableMmr have been updated to include a
Bitmap in the constructor.

The LMDB_DB constructor can now load the deleted UTXO state
from the backend and populate the UTXO Mmr instance. This
is Step 1 of making the UTXO state persist which UTXOs have
been spent.

The next step (not done here) is to write the UTXO bitmap to
the backend when it changes.