Simultaneous connection tie breaking
If peers dial each other at the same time there must be a way for both
peers to deterministically choose which connection to keep and which to
disconnect.

This PR adds code to handle this situation by using the NodeID and
ConnectionDirection to enable both sides to agree on a connection to
close.