Resolve issue #1132
As per the discussion in #1132, the failing test, `check_pruning_of_pruned_mmr`
is a case of the expected behaviour of the class and how it was designed
to work being at odds.

Part of this confusion stems from ChangeTracker not having a Single
Responsiblity.

Therefore, we've decided to proceed as follows:

* Blockchain data synchronisation does not need to have anything to do
with a ChangeTracker if the CheckPoint data (the additions / deletions
per block) are promoted to first-class citizens in the backing database.
* Then the moving window history that requires the `base` mmr field to
be mutable can be removed. the `base` MMR can then be immutable as
originally intended.
* The ChangeTracker's resposibility is thus i) keeping track of the
chain tip MMR root. ii) Holding state necessary for perorming re-orgs.

Whoops, It still has multiple responsibilities!

Ok, a single PrunableMMR stored in memory can be used to keep track of
the chain tip MMR root. The databse backend implementation can choose to
periodically "re-base" the in-memory instance from time to time so that
the data structure does not grow indefinitely.

And rewind logic is a matter of truncating the checkpoint storage and
rebasing the tip's MMR instance.

Goodbye ChangeTracker, it was nice knowing you. Goodbye #1132