Add testnet faucet generator
To help wallet users get some testnet Tari to play with, we're
adding some utxos to the genesis block that they can claim
from their mobile wallet apps.

This PR add a small application that generates the faucet UTXOs
and private keys.

The UTXOs will be added to the alpha-net genesis block.

The `create_utxo` helper function was extended to accept `OutputFeatures`