Add checksum algorithm
The luhn mod 1024 checksum is added to allow us to detect transposition
errors in emoji ids. Luhn was selected over e.g. Damm becuase the Damm
algorithm was developed for 10-digit "alphabets", whereas we have a 1024
digit alphabet.

The weaknesses of the Luhn algorithm (missing some errors) are mitigated
as the alphabet size increases, so the false positive rate for emoji ids
should be neglibly small.