Don't redial peer if we cannot connect to them
Currently, if we cannot send to a peer the user of `MessagingProtocol`
has no way to know if the peer is not contactable.

This PR notifies the requester if they attempting to dial a peer which was not
contactable within a cooldown period (1 minute) so that they have some
way to handle that (give up, retry later etc).

The messaging protocol provides the following reasons as to why a message
failed to send:
```rust
pub enum SendFailReason {
    /// Dial was not attempted because the peer is offline
    PeerOffline,
    /// Dial was attempted, but failed
    PeerDialFailed,
    /// Outbound message envelope failed to serialize
    EnvelopeFailedToSerialize,
    /// Failed to open a messaging substream to peer
    SubstreamOpenFailed,
    /// Failed to send on substream channel
    SubstreamSendFailed,
}
```