Dry up node builder
* The node builder has been cleaned up significantly.

The `BlockchainBackend` has been abstracted away behind a
`NodeContainer` enum that exposes a set of methods that
are backend-agnostic.

This lets us remove all the duplicated code in the builder
logic.

All the functions in the builder have been refactored to have
a *single responsibility*.

The miner builder code has been collected up and grouped together.

* A hash rate calculator hash been added to `blake_miner`.

Every minute, if a block hasn't been found, it'll log the hashrate to
the logs (in MH/s).

* Many log messages have been added to the code, the mining code in
particular.

* Add trace logs to log the difficulty calculations

* Rename testnet => Rincewind

Tari's first testnet is named Rincewind, the first major character from
the Discworld series; The code base had some references to it here and
there, this commit changes references to the config file as well so that
naming is consistent throughout.