Add configuration file JSON definition
Add the JSON config definition that Tari Config Generator will use to
create the configuration options.

Add a README explaining the purpose of the file.

Create the presets folder tht will hold a set of preset config files.

Update the loadup prompt to indicate the presence of the config
generator