Update FFI Tx listing functions for pending & completed transactions
The frontend considers a transaction as Completed only once it is mined. However the Transaction manager considers a CompletedTransaction to be one that have been completely negotiated between parties and is now ready to be Broadcast and Mined.

In order to present both truly pending and CompletedTransactions in the Completed and Broadcast states as Pending to the frontend this PR adds some logic to return them in the `wallet_get_pending_outbound_transactions(…)` and `wallet_get_pending_inbound_transactions(…)` FFI function calls. They are also excluded from the `wallet_get_completed_transactions(…)` function call which only include CompletedTransactions with the Mined status.