Added supported protocols to Peer
Add supported protocols for each peer. This is obtained/updated when a peer connection
is established as part of the identity exchange.

This list will be used to allow optimistic protocol negotiation. That
is, instead of waiting for a peer to confirm if they speak a particular
protocol, that can be safely assumed if the protocol is in the
`supported_protocols` list. Thereby, eliminating the need to wait for a
protocol confirmation message and allowing the protocol to begin
immediately. Optimistic protocol negotiation is TODO.