Fix for mobile wallet discovery
The DHT did not propagate private discovery messages to wallets because
they do not have the `MESSAGE_PROPAGATION` feature.

This PR changes the behaviour of the DHT to include wallets in
discovery propagation messages. Internally, the Forward step of the DHT pipeline
has a special case that will propagate to wallets. The closest peer selection
has been adjusted to allow the caller to specify the PeerFeatures of the closest nodes to target by using the `neighbours_with_features` call.
_NOTE:_ because those nodes will not propagate the message further, the selection ensures
that a minimum number of propagation nodes are always selected

This PR also includes a new example called "memorynet". `memorynet`, as
the name suggests, is a small in-memory DHT network consisting of a
number of "base nodes" and "wallets" (all that differs is their feature
flags).

The following happens when running `memorynet`:
1. A single "seed node", `NUM_NODES` "base nodes" and `NUM_WALLETS` "wallets" are generated and started.
1. All "base nodes" join the network via the "seed node"
1. All "wallets" join the network via a random "base node"
1. The first "wallet" in the list attempts to discover the last "wallet" in the list
1. The result of the discovery is printed and the network is shut down

The example can be run with this command:
`RUST_BACKTRACE=1 RUST_LOG=trace cargo run --example memorynet 2> /tmp/debug.log`