Fixed "missing peers" LMDB issue
This PR seems to solved the "missing peers" conundrum.

Background: using the async peer manager (blocking threadpool, sync
rwlock), a peer would be added in one component, soon afterwards another
component would not be able to find that peer. This did not happen when
changing to the sync peer manager. Thanks to @neonknight and
@blackwolfsa who provided a test which reliably reproduces this issue.

Here are my observations:

1. This did not happen when using the in-memory peer db backend (presumably
that eliminates the possibility that there is a code error in peer
manager)
1. Internal peer manager indexes contain the correct peer key references, however lmdb
   `get` reports not to have the related peer at the given key.
1. This may be related to inserting/overwriting a peer which already exists
(i.e. when updating a peer)
1. An exclusive sync mutex made no difference

The fix: Use tokio's async RwLock and make the peer manager interface
fully async, do away with the blocking thread implementation