Implemented optimistic protocol negotiation
This PR removes a round of comms when negotiating protocols if the protocols supported
by the peer are known ahead of time (see #1404).

The "normal" negotiation protocol as follows (best case):
1. Initiator sends protocol candidate, and waits
1. Responder responds with TERMINATE flag and name of protocol to speak
1. Protocol begins

The "optimistic" negotiation is as follows:
1. Initiator sends protocol candidate (no waiting)
1. Protocol begins

Responder will receive the protocol name with the OPTIMISTIC flag and
either close the stream (if invalid) or accept and begin the protocol.