Move waiting_requests to base node
Code under the `mempool_proto` feature required `RequestKey` among
others. The latter is currently in the helpers module, which is under
the `base_node` feature flag.

In addition, the helpers module is documented as holding functions for
making tests easier, which suggests that we should move waiting_requests
rather than have fine-grained flags on the helper module.

This commit merely moves waiting_requets to the base_node module and
applies the relevant *_proto feature flags to it.

Also removed some unused imports