Add quick and dirty OsX installer
This PR is a minimal viable installer for OsX.

The create_bundle script packages the release binary with working
minimal config files.
Then install-osx will install dependencies, create an id and run the
node. It also runs Tor in a separate terminal.

As I said, this is an MVP. There are lots of enhancements that can and
should be made to this; but this is a start.