Detect failure and reestablish tor control connection + tor events
If the tor server shuts down (for whatever reason), the hidden service
would cease to exist on the tor network making it impossible for clients
to connect into the node until the node is restarted.

This PR fixes this by adding a controller that is responsible for tor
hidden service reliability. The tor controller will attempt
to reconnect every 5 seconds until either it succeeds or is shut down.

Additionally, the TorControlClient now has the ability to subscribe to
events emitted by the tor control server and publish those on a tokio
broadcast channel. This could be useful down-the-line for determining
network availability/bandwidth usage etc.

This implementation required that:
1. The tor control connection is monitored in it's own task
1. Events and response text is split into seperate streams
1. The hidden service is immediately recovered once a connection is
   reestablished.
1. The event stream is preserved when reestablishing connections to tor
1. The SETEVENTS command is supported