Remove R2D2 for Sqlite from Wallet
R2D2 was added to provide thread safety for the Async wallet services with the thought that in the future we might require a pool of connections but after testing we are having problems with Set faults on android coming from the R2D2 thread pool.

So this PR removes the threadpool in favour of a single SqliteConnection maintained behind a Arc<Mutex<>>. This is a much simpler arrangement and will hopefully be more stable on mobile platforms.