Ensure yamux incoming worker stops if connection closed impolitely
If the underlying socket was suddenly dropped on the floor by a peer
a stream of eof error messages would flood the logs. It appears as
if the yamux stream does not close immediately in this case but insteads
emits a ConnectionError:Closed. This PR will stop the inbound worker in
this case.