Base node also starts a wallet comms node
The base node now starts an independent wallet comms node. Mined coinbase
UTXOs are sent to this wallet.

Updated the cli commands appropriately.

I attempted to cross the wires of the base node and wallet services
messaging pipelines but this ended up being a little too messy and complex.
Some of the code in this PR are some generally useful changes from this attempt.

Did some general cleanup and cleanup of the "path as a string" anti-pattern.