Update to tari_utils 0.1.2 and send_tari arg parsing
The send_tari command was causing a thread panic if an emoji id was
used. In refactoring the parsing logic, it was discovered that the
problem lay with `from_hex` in the tari_utilities crate.

The crate was patched with a fix and bumped to 0.1.2.

The send_tari command now works with both hex and emoji pubkeys.