Fix for UTXO sync in output manager
There was a bug in the Output Manager Service UTXO Sync process where if a UTXO was imported into the OMS after a Sync message was sent but before the response was received.

In this case the hash of the imported UTXO was not sent in the query and when the resposne came it does not include the new UTXO which was considered to be invalid.

This PR adds the list queries UTXO hashes to the list of pending sync queries and only hashes that are in that list and not in the response and invalidated.