If tor socket has some kind of error, it shouldn't be read from again
The contract for AsyncRead is generally that if there is an `Some(Err(_))`
or `Some(None)`, `poll_read` should never be called again. This fix
breaks the loop if a socket error is received. The tor controller will
attempt to reconnect.

This prevents infinite looping of this log (confirmed on iOS app):
```
 [comms::tor::control_client] ERROR Line framing error when reading from tor control server: 'Io(Os { code: 57, kind: NotConnected, message: "Socket is not connected"  })'
```

The following is unclear:
1. The reason this happens in the first place (tor crash?)
1. If the connection can be reestablished (maybe tor crashed and
   restarted? If so then reconnect will work)