Update transaction service to monitor for completed tx’s being mined
Currently the Transaction service monitors for CompletedTransactions being Broadcast and then after that transition for Broadcast transactions to change to being Mined. However if for some reason the Transaction service missed a transaction being broadcast and it moves straight to being mined the Transaction service wouldn’t note that. 

This PR updates the transaction service to also check if Completed Transactions are mined and not just broadcast to cater for this case.