Fix bug when retrying messsage send in MessagingProtocol
Messaging protocol makes a contract with clients that for every message
sent, you will either get back a MessageSent or SendMessageFailed event
for that (tagged) message.

Messaging protocol did not add the message to the retry queue after
one attempt. As a side effect of the bug, the above contract was broken.