Tidy up test_event_channel
This test is supposed to test the event channel of the node state
machine but in doing so it originally spun up a lot of
infrastructure
including 2 complete base nodes and a communication stack.

The test has been refactored to make use of a new ChainMetadata Mock
structure which
allows us to inject chain metadata messages directly to the state
machine without having to go the circuitous route of spinning up a
peer
node and constructing a blockchain.