Update Transaction Service to handle Discovery and Failed sends properly
The behaviour in the transaction service when sending transaction where discovery failed or when discovery passed but the actually sending failed was wrong. It results in a pending transaction being added to the wallet that could never be completed as it was never sent and will never be resent. Another issue was that if during discovery the app was closed or crashed the encumbered UTXOs will never be released. The idea of short-term encumbrance was added to the Output Manager service to deal with this.

This PR implements the following:
- The handling of the flow of sending a transaction with the various permutation of discovery and send failure are now handled correctly
- Found a bug in the MessageEvent handling in the Comms layer that signals when a transaction was actually sent and now handles it correctly.
- Added a new short term encumbrance feature to the Output Manager Service to allow for emcumbering UTXO’s while discovery is in progress but if the app is turned off or crashes the encumbrance will be cleared on startup if the transaction did not complete.