Add waiting state to base node
If a node is lagging but unable to synchronise with the network (for any
reason), it will rapidly oscillate between `Listening` and `BlockSync`.

This is not necessarily the wrong behaviour; it's just the speed that is
the problem.

This PR adds a new `Waiting` state that puts the node into a time-out
when block synchronisation fails before going back into Listening state.

I've moved the functions from the "helper" module back to where they
were originally. The reasons are:
i) They're not used anywhere else, so I can make them private.
ii) You can change these functions in future confident in the knowledge
that you're not breaking code somewhere else because they're not public
functions.