Added configurable liveness check mode to wire protocol
Allow the initiator to select a wire protocol by sending a wire protocol
byte. `0x01` is the normal comms protocol, `0x45` ("E") is an echo
service protocol that can be used for TCP liveness. This protocol is
extensible so that other modes can be added if ever required without a
hard network fork.

Added `liveness_max_sessions` configuration. When set to 0, no liveness
sessions are permitted (default). For those wishing to monitor liveness
of their base node, this can be set to 1 to allow a single monitor to
connect and monitor liveness. Any subsequent connections that initiate a
liveness session will be closed immediately until the monitor
disconnects. CIDRs must be specified to restrict the source of the
connection (e.g. to localhost only `127.0.0.1/32`).

Care has been taken to ensure that this mode cannot be exploited in some
way as data sent over this connection is unauthenticated and unencrypted.

The following restrictions apply:
- The number of sessions must not exceed those specified in the
configuration
- The peer address must be contained the given CIDR range.
- The maximum length of a line sent is 50
- The initiator has 7 seconds to send the first byte

If any of these are not satisfied, the connection is terminated.

_NOTE: hard network fork. Out of date and updated base nodes will not be able to communicate_