[Windows] Exclude non-critical DHT messaging layers
Windows `__chkstk` routine throws an access violation when all DHT layers are
added. It seems not to be the size of the stack, but the depth. This is
not fully understood yet. As a work around, this PR reintroduces the store
and forward layers excluded previously and excludes layers that are not
critical to a properly functioning node (namely, MessageLogging and
the filtering out of store and forward messages for mobile wallets)