Add local mempool service struct
Provide a LocalMempool Service struct that acts as a wrapper around a
Sender-Requester pair.

A single utility function `request_stats` is added that handles the
channel mangling for you and returns a future for the current  mempool stats.

This actual stats colleciton is not done right now; a subsequent PR will
connect the request stream to the main mempool service loop.

A test that demonstrates that the service is functional is provided.