Edge case for message propagation, DHT logging improvements
Handled the case where a peer cannot decrypt a message destined for
themselves by:
1. Not propagating the message that could not be decrypted in this case
1. Discarding a message that was forwarded by a node that could not
   decrypt the message

It is not clear how a message that could not be decrypted by the
destination entered the (currently non-adversarial) network, but
this case needed to be handled in any case.

Also added store and forward test in memorynet.
TODO: It's clear that there are way too many messages entering store and
forward. Some kind of S+F subscription protocol would be ideal, for now
some improvements on prioritising which messages should be stored and
returned from S+F requests should prove worthwhile.