Fix block period calculation
The period calculation had an error in the minimum calculation.
This PR rewrites that calculation:
a) putting the calculation itself in a helper function.
b) using functional programming to make the logic more expressive.
c) adding a few unit tests to confirm the logic is correct (in at least one
case)