Update weights
The existing transaction weights were placeholder values.

This PR updates the weights to be more representative of their cost in
disk and bandwidth usage.

A single input is 57 bytes, and output is 737 B and a kernel is 147 B.
Normalising an input weight to one, an output would be 12.9 and a kernel 2.6.

There is one kernel per transaction; and one header per block. We can't
know what the header contribution is per transaction, but by rounding
the kernel weight to 3, each transaction can "pay" for a little part of
the header.

Rounding the output weight to 13 makes very little difference to
the overall arithmetic.

Overall, this makes every transaction, regardless of number of in- or
outputs roughly the same weight per byte; around 57 ±  bytes/weight.

Several tests were updated that used fees in theor assertions to account
for the new weights; typically by reducing the fee per gram specified in
the test transactions.