Calcuate region threshold depending on PeerFeatures
When a node is determining if a requesting node is in it's region, it takes
the closest `n` peers that it is aware of and if the node ID is within
that threshold it considers it a neighbouring node. Because there are
nodes on the network that do not perform some network functions (such as
message propagation and store and forward i.e. wallets), there is a
chance that a majority of neighbours are these non-propagating nodes and
'take up space' where propagation nodes could 'fit'.

When calculating the threshold for neighbouring nodes it must take the
features of the nodes it knows about into account.

This PR changes the peer manager to consider `COMMUNICATION_NODE`s and
`COMMUNICATION_CLIENT`s independently to ensure a minimum number of
propagation nodes are considered neighbouring.