Modular configuration via ConfigLoader and ConfigPath traits
This PR adds a way to derive configuration loading logic in a tari-way via
ConfigLoader and ConfigPath traits. Config crate plays nicely with Serialize
and Deserialize traits, this approach is reusing serde power as much as
possible to make defining and loading configuration structs as easy as possible.
It also provides default parameters setting reusing Default trait.

As part of this change module documentation was extended and configuration.rs file
was split on submodules. It also deprecates ConfigExtractor trait:

- add ConfigLoader and ConfigPath traits with examples
- add DefaultConfigLoader and NetworkConfigPath traits with examples
- move bootstrap impl from lib to configuration/bootstrap.rs submodule
- improve documentation for main lib and submodules
- move GlobalConfig and utilities to sub modules under configuration/

- apply code review fixes