Update wallet to support sending Tx via Store and Forward
This PR adds support for sending transaction via Store and Forward to the Wallet.

When a transaction is sent it is now attempted to send a Direct message which includes a Discovery attempt but also sends the message to the network to be propagated so that the recipient will be able to receive it via Store and Forward when they come back online. The wallet is updated to tolerate receiving repeated messages in case both these messages arrive.

As long as a transaction was sent out in either channel (direct or via store and forward) the transaction will be considered a Pending Outbound Transaction and will start waiting for a Transaction Reply.

Because Store and Forward is more async and fire and forget this means that if too much time has passed the transaction will need to be explicitly cancelled. For this a new function is provided: `transaction_service.cancel_transaction(tx_id: TxId)`. This will cancel any Pending Outbound Transaction. An FFI function is provided for this functionality.

Two new callbacks are provided for FFI clients to interface with the new flow:
- `callback_direct_send_result: unsafe extern "C" fn(TxId, bool)`: Will return the result of the attempt at a Direct send of a Tx
- `callback_store_and_forward_send_result: unsafe extern "C" fn(TxId, bool)`: Will return the result of trying to send a tx out via the network

One final change introduced in this PR is a new callback to indicate when a transaction has been cancelled due to a Blockchain Monitoring event. For example if the mempool rejected the transaction.
- `callback_transaction_cancellation: unsafe extern "C" fn(TxId)`: Will be called when a transaction is cancelled (including when it is manually cancelled by the FFI client)

*The `callback_discovery_process_complete` is removed in this PR*

## Summary:
- Store and forward implemented
- callback_direct_send_result added
- callback_store_and_forward_send_result added
- callback_transaction_cancellation added
- callback_discovery_process_complete removed